% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emu_function.R
\name{emu_function}
\alias{emu_function}
\title{A function about emus}
\usage{
emu_function(love = T)
}
\arguments{
\item{love}{Do you love emus? Defaults to TRUE}
}
\description{
What bird should you hug today? Find out here
}
\examples{
emu_function()
}
\keyword{bird}
